+++ infra/modules/aws_azure_vpn/aws_data.tf
@@ -1,0 +1,19 @@
+data "aws_region" "current" {}
+data "aws_caller_identity" "current" {}
+
+# Latest Amazon Linux 2023 AMI
+data "aws_ami" "amazon_linux" {
+  count       = var.use_case == "development" ? 1 : 0
+  most_recent = true
+  owners      = ["amazon"]
+
+  filter {
+    name   = "name"
+    values = ["al2023-ami-2023.8.20250915.0-kernel-6.1-x86_64"]
+  }
+
+  filter {
+    name   = "virtualization-type"
+    values = ["hvm"]
+  }
+}