+++ infra/resources/_modules/testing/variables.tf
@@ -1,0 +1,30 @@
+variable "test_modes" {
+  type        = set(string)
+  description = "List of test kinds to create resources for. Allowed values are 'integration' and 'e2e'."
+
+  validation {
+    condition     = alltrue([for v in var.test_modes : contains(["integration", "e2e"], v)])
+    error_message = "Invalid value in test_modes. Allowed values are 'integration' and 'e2e'."
+  }
+}
+
+variable "environment" {
+  type = object({
+    prefix          = string
+    environment     = string
+    location        = string
+    instance_number = string
+  })
+}
+
+variable "tags" {
+  type        = map(string)
+  description = "Tags to apply to resources"
+}
+
+variable "gh_pat_reference" {
+  type = object({
+    keyvault_name                = string
+    keyvault_resource_group_name = string
+  })
+}