+++ infra/modules/aws_github_environment_bootstrap/CHANGELOG.md
@@ -1,5 +1,11 @@
 # aws_github_environment_bootstrap
 
+## 0.0.3
+
+### Patch Changes
+
+- 11ef401: Include oidc_provider_arn variable in examples and documentation
+
 ## 0.0.2
 
 ### Patch Changes